Abendliches gepfusche, Unhandled Exaption

+++ os/various/MPU6050/helper_3dmath.h
@@ -31,7 +31,7 @@ THE SOFTWARE.
 
 #ifndef _HELPER_3DMATH_H_
 #define _HELPER_3DMATH_H_
-
+#include <math.h>
 class Quaternion {
     public:
         float w;
@@ -108,7 +108,7 @@ class VectorInt16 {
         }
 
         float getMagnitude() {
-            return sqrt(x*x + y*y + z*z);
+            return sqrtf(x*x+y*y+z*z);
         }
 
         void normalize() {
